Updated docs, changelog and expanded error handling
