+++ Sources/Paywall/Documentation.docc/GettingStarted.md
@@ -4,7 +4,7 @@ Configuring the SDK and its delegate.
 
 ## Overview
 
-To get up and running, you need to get your **API Key** from the Superwall Dashboard. You then configure the SDK using ``Paywall/Paywall/configure(apiKey:userId:delegate:options:)``. You pass this function a class that conforms to ``PaywallDelegate``, which handles actions taken on the paywall. You can then present your paywall.
+To get up and running, you need to get your **API Key** from the Superwall Dashboard. You then configure the SDK using ``Paywall/Paywall/configure(apiKey:delegate:options:)``. You pass this function a class that conforms to ``PaywallDelegate``, which handles actions taken on the paywall. You can then present your paywall.
 
 ## Getting your API Key
 
@@ -17,7 +17,7 @@ On that page, you will see your **Public API Key**. Copy this for the next step.
 
 ### Configuring the SDK
 
-To configure the SDK, you call ``Paywall/Paywall/configure(apiKey:userId:delegate:options:)`` from  `application(_:didFinishLaunchingWithOptions:)`. We recommended creating a service class **PaywallService.swift** that handles your SDK configuration and delegate callbacks:
+To configure the SDK, you call ``Paywall/Paywall/configure(apiKey:delegate:options:)`` from  `application(_:didFinishLaunchingWithOptions:)`. We recommended creating a service class **PaywallService.swift** that handles your SDK configuration and delegate callbacks:
 
 ```swift
 import Paywall
@@ -29,7 +29,6 @@ final class PaywallService {
   static func initPaywall() {
     Paywall.configure(
       apiKey: apiKey,
-      userId: "aUniqueUserId"
       delegate: shared
     )
   }
@@ -38,9 +37,14 @@ final class PaywallService {
 
 This configures a shared instance of ``Paywall/Paywall`` for use throughout your app and sets the delegate to the `PaywallService` shared instance. Make sure to replace the `apiKey` with your Public API key that you just retrieved.
 
-If you have a unique ID of your user, you can enter it in `userId`. Otherwise we will generate a random id that will persist internally until you call ``Paywall/Paywall/reset()`` or the user deletes/reinstalls your app.
 
-Later on, you can identify a user using ``Paywall/Paywall/identify(userId:)``. Superwall will then alias the userId you provide to the random userId we generated on your behalf.
+## Identity Management
+
+We generate a random user ID that persists internally until ``Paywall/Paywall/logOut()`` or ``Paywall/Paywall/reset()`` is called or the user deletes/reinstalls your app.
+
+However, if you use your own user management system, you can call ``Paywall/Paywall/createAccount(userId:)`` when a user first creates an account, and ``Paywall/Paywall/logIn(userId:)`` if you're logging in an existing user. This will alias your `userId` with the anonymous Superwall ID enabling us to load the user's assigned paywalls.
+
+Calling ``Paywall/Paywall/logOut()`` will reset the on-device userId to a random ID and clear the paywall assignments.
 
 ## Setting the delegate
 
@@ -56,8 +60,7 @@ extension PaywallService: PaywallDelegate {
 
   // 2
   func restorePurchases(completion: @escaping (Bool) -> Void) {
-    // TODO: Restore purchases and call completion block with boolean indicating
-    // the success status of restoration
+    // TODO: Restore purchases and call completion block with boolean indicating the success status of restoration
     completion(false)
   }
 
@@ -74,6 +77,7 @@ These delegate methods are mandatory and receive callbacks from the SDK in respo
 1. Purchasing a given product. In here, enter your code that you use to purchase a product.
 2. Restoring purchases. Make sure to call the completion block after you attempt to restore purchases to let the SDK know whether the restoration was successful or not.
 3. Telling the SDK whether the user has an active subscription. Replace this with a boolean indicating the user's subscription status.
+
 
 ## Configuring From the App Delegate
