+++ Sources/Paywall/Paywall/Game Controller/GameControllerManager.swift
@@ -4,6 +4,7 @@
 //
 //  Created by Jake Mor on 10/5/21.
 //
+// swiftlint:disable force_unwrapping
 
 import Foundation
 import GameController
@@ -13,6 +14,7 @@ protocol GameControllerDelegate: NSObject {
 	func gameControllerEventDidOccur(event: GameControllerEvent)
 }
 
+@MainActor
 final class GameControllerManager: NSObject {
 	static let shared = GameControllerManager()
 	weak var delegate: GameControllerDelegate?
@@ -21,19 +23,14 @@ final class GameControllerManager: NSObject {
     _ name: String,
     _ value: Float
   ) {
-		DispatchQueue.main.async { [weak self] in
-			guard let self = self else {
-        return
-      }
-			let event = GameControllerEvent(
-        controllerElement: name,
-        value: Double(value),
-        x: 0,
-        y: 0,
-        directional: false
-      )
-			self.delegate?.gameControllerEventDidOccur(event: event)
-		}
+    let event = GameControllerEvent(
+      controllerElement: name,
+      value: Double(value),
+      x: 0,
+      y: 0,
+      directional: false
+    )
+    self.delegate?.gameControllerEventDidOccur(event: event)
 	}
 
 	func valueChanged(
@@ -41,19 +38,14 @@ final class GameControllerManager: NSObject {
     _ x: Float,
     _ y: Float
   ) {
-		DispatchQueue.main.async { [weak self] in
-			guard let self = self else {
-        return
-      }
-			let event = GameControllerEvent(
-        controllerElement: name,
-        value: 0,
-        x: Double(x),
-        y: Double(y),
-        directional: true
-      )
-			self.delegate?.gameControllerEventDidOccur(event: event)
-		}
+    let event = GameControllerEvent(
+      controllerElement: name,
+      value: 0,
+      x: Double(x),
+      y: Double(y),
+      directional: true
+    )
+    self.delegate?.gameControllerEventDidOccur(event: event)
 	}
 
   func gamepadValueChanged(
@@ -77,12 +69,10 @@ final class GameControllerManager: NSObject {
     case gamepad.leftThumbstick:
       self.valueChanged(name, gamepad.leftThumbstick.xAxis.value, gamepad.leftThumbstick.yAxis.value)
     case gamepad.leftThumbstickButton:
-      // swiftlint:disable:next force_unwrapping
       self.valueChanged(name, gamepad.leftThumbstickButton!.value)
     case gamepad.rightThumbstick:
       self.valueChanged(name, gamepad.rightThumbstick.xAxis.value, gamepad.rightThumbstick.yAxis.value)
     case gamepad.rightThumbstickButton:
-      // swiftlint:disable:next force_unwrapping
       self.valueChanged(name, gamepad.rightThumbstickButton!.value)
     case gamepad.dpad:
       self.valueChanged(name, gamepad.dpad.xAxis.value, gamepad.dpad.yAxis.value)
@@ -117,54 +107,4 @@ final class GameControllerManager: NSObject {
       )
     }
 	}
-}
-
-extension GCControllerElement {
-  // the identifier of this gamecontroller element that is accepted by the paywall javascript event listeners
-  @available(iOS 13, macOS 10.15, *)
-  func buttonName(gamepad: GCExtendedGamepad) -> String {
-    switch self {
-    case gamepad.leftTrigger:
-      return "L2 Button"
-    case gamepad.leftShoulder:
-      return "L1 Button"
-    case gamepad.rightTrigger:
-      return "R2 Button"
-    case gamepad.rightShoulder:
-      return "R1 Button"
-    case gamepad.leftThumbstick:
-      return "Left Thumbstick"
-    case gamepad.leftThumbstickButton:
-      return "Left Thumbstick Button"
-    case gamepad.rightThumbstick:
-      return "Right Thumbstick"
-    case gamepad.rightThumbstickButton:
-      return "Right Thumbstick Button"
-    case gamepad.dpad:
-      return "Direction Pad"
-    case gamepad.dpad.down:
-      return "Direction Pad"
-    case gamepad.dpad.right:
-      return "Direction Pad"
-    case gamepad.dpad.up:
-      return "Direction Pad"
-    case gamepad.dpad.left:
-      return "Direction Pad"
-    case gamepad.buttonA:
-      return "A Button"
-    case gamepad.buttonB:
-      return "B Button"
-    case gamepad.buttonX:
-      return "X Button"
-    case gamepad.buttonY:
-      return "Y Button"
-    case gamepad.buttonMenu:
-      return "Menu Button"
-    case gamepad.buttonOptions:
-      return "Options Button"
-    default:
-      // Logger.debug(logLevel: .debug, scope: .gameControllerManager, message: "Unrecognized Button", info: ["button": element], error: nil)
-      return "Unknown Button"
-    }
-  }
 }
