+++ Sources/Paywall/Products/In-App Receipt/InAppReceiptAttribute.swift
